Introduce MergeOptionsBase
Introduce a MergeOptionsBase and a MergeAndCheckoutOptionsBase so
that there's less duplication of the various options that go to
merge-like things.
