+++ LibGit2Sharp/MergeOptions.cs
@@ -7,7 +7,7 @@ namespace LibGit2Sharp
     /// <summary>
     /// Options controlling Merge behavior.
     /// </summary>
-    public sealed class MergeOptions : IConvertableToGitCheckoutOpts
+    public sealed class MergeOptions : MergeAndCheckoutOptionsBase
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="MergeOptions"/> class.
@@ -20,88 +20,12 @@ namespace LibGit2Sharp
         /// </summary>
         public MergeOptions()
         {
-            CommitOnSuccess = true;
-
-            FindRenames = true;
-            // TODO: libgit2 should provide reasonable defaults for these
-            //       values, but it currently does not.
-            RenameThreshold = 50;
-            TargetLimit = 200;
         }
-
-        /// <summary>
-        /// The Flags specifying what conditions are
-        /// reported through the OnCheckoutNotify delegate.
-        /// </summary>
-        public CheckoutNotifyFlags CheckoutNotifyFlags { get; set; }
-
-        /// <summary>
-        /// Commit the merge if the merge is successful and this is a non-fast-forward merge.
-        /// If this is a fast-forward merge, then there is no merge commit and this option
-        /// will not affect the merge.
-        /// </summary>
-        public bool CommitOnSuccess { get; set; }
 
         /// <summary>
         /// The type of merge to perform.
         /// </summary>
         public FastForwardStrategy FastForwardStrategy { get; set; }
-
-        /// <summary>
-        /// How conflicting index entries should be written out during checkout.
-        /// </summary>
-        public CheckoutFileConflictStrategy FileConflictStrategy { get; set; }
-
-        /// <summary>
-        /// Find renames. Default is true.
-        /// </summary>
-        public bool FindRenames { get; set; }
-
-        /// <summary>
-        /// Similarity to consider a file renamed.
-        /// </summary>
-        public int RenameThreshold;
-
-        /// <summary>
-        /// Maximum similarity sources to examine (overrides
-        /// 'merge.renameLimit' config (default 200)
-        /// </summary>
-        public int TargetLimit;
-
-        /// <summary>
-        /// How to handle conflicts encountered during a merge.
-        /// </summary>
-        public MergeFileFavor MergeFileFavor { get; set; }
-
-        /// <summary>
-        /// Delegate that the checkout will report progress through.
-        /// </summary>
-        public CheckoutProgressHandler OnCheckoutProgress { get; set; }
-
-        /// <summary>
-        /// Delegate that checkout will notify callers of
-        /// certain conditions. The conditions that are reported is
-        /// controlled with the CheckoutNotifyFlags property.
-        /// </summary>
-        public CheckoutNotifyHandler OnCheckoutNotify { get; set; }
-
-        #region IConvertableToGitCheckoutOpts
-
-        CheckoutCallbacks IConvertableToGitCheckoutOpts.GenerateCallbacks()
-        {
-            return CheckoutCallbacks.From(OnCheckoutProgress, OnCheckoutNotify);
-        }
-
-        CheckoutStrategy IConvertableToGitCheckoutOpts.CheckoutStrategy
-        {
-            get
-            {
-                return CheckoutStrategy.GIT_CHECKOUT_SAFE |
-                       GitCheckoutOptsWrapper.CheckoutStrategyFromFileConflictStrategy(FileConflictStrategy);
-            }
-        }
-
-        #endregion
     }
 
     /// <summary>
